improvement: block empty keys
In this commit we block empty keys from inputs

Empty key can be either a direct query with ""
or a path query e.g. str1..str2.str3

In the above example, after str1 we have a double
dot. The methods currently split keys with dot
delimeter before a search is initiated.

Note: While the algorithm works with dots also,
it makes no sense to allow empty keys as part
of map content