improvement: [SQL] switch to pointers
In this commit we are switching input/return interfaces
to pointer interfaces. For bigger volume of data, this
change will be helpful to optimize the code to do as less
copies as possible.

For now we are not changing the logic of sql functions but
only update them to handle pointers

Breaking change:
- SQL methods will return pointers.
  Only Storage.Wrappers will handle this change automatically